proptypes, move to modules
