+++ src/modules/swiper/Card/styles.js
